Converge Container Start/Stop/Run/Create into CLI model (#14240)

+++ src/windows/wslc/services/ContainerModel.h
@@ -17,8 +17,10 @@ Abstract:
 #include <wslservice.h>
 #include <wslaservice.h>
 #include <docker_schema.h>
+#include <string>
 
 namespace wsl::windows::wslc::models {
+
 struct ContainerCreateOptions
 {
     bool TTY = false;
@@ -41,7 +43,7 @@ struct StopContainerOptions
 {
     static constexpr LONGLONG DefaultTimeout = -1;
 
-    int Signal = WSLASignalSIGTERM;
+    ULONG Signal = WSLASignalSIGTERM;
     LONGLONG Timeout = DefaultTimeout;
 };
